opt stcp/build
