+++ .github/workflows/goreleaser.yml
@@ -1,0 +1,33 @@
+name: goreleaser
+
+on:
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  goreleaser:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Set up Go
+        uses: actions/setup-go@v3
+        with:
+          go-version: 1.19
+
+      - name: Make All
+        run: |
+          ./package.sh
+      - name: Run GoReleaser
+        uses: goreleaser/goreleaser-action@v3
+        with:
+          version: latest
+          args: release --rm-dist
+          # workdir: ./cmd/frps_allowed_ports
+        env:
+          GITHUB_TOKEN: ${{ secrets.GPR_TOKEN }}
