--- .travis.yml
@@ -1,9 +1,0 @@
-sudo: false
-language: go
-
-go:
-    - 1.13.x
-    - 1.14.x
-
-install:
-    - make
