+++ Makefile.cross-compiles
@@ -1,0 +1,22 @@
+export PATH := $(GOPATH)/bin:$(PATH)
+export GO111MODULE=on
+LDFLAGS := -s -w
+
+os-archs=darwin:amd64 darwin:arm64 linux:386 linux:amd64 linux:arm linux:arm64 windows:386 windows:amd64
+
+all: build
+
+build: app
+
+app:
+	@$(foreach n, $(os-archs),\
+		os=$(shell echo "$(n)" | cut -d : -f 1);\
+		arch=$(shell echo "$(n)" | cut -d : -f 2);\
+		gomips=$(shell echo "$(n)" | cut -d : -f 3);\
+		target_suffix=$${os}_$${arch};\
+		echo "Build $${os}-$${arch}...";\
+		env CGO_ENABLED=0 GOOS=$${os} GOARCH=$${arch} GOMIPS=$${gomips} go build -trimpath -ldflags "$(LDFLAGS)" -o ./release/frps_allowed_ports_$${target_suffix} ./cmd/frps_allowed_ports;\
+		echo "Build $${os}-$${arch} done";\
+	)
+	@mv ./release/frps_allowed_ports_windows_386 ./release/frps_allowed_ports_windows_386.exe
+	@mv ./release/frps_allowed_ports_windows_amd64 ./release/frps_allowed_ports_windows_amd64.exe
