--- cmd/frp_plugin_allowed_ports/main.go
@@ -1,4 +1,0 @@
-package main
-func main() {
-	Execute()
-}
