--- cmd/frp_plugin_allowed_ports/ports
@@ -1,6 +1,0 @@
-user1=6000
-user1=321
-user1=6262
-user1=service
-user1=service.masternetwork.us
-user2=124
