+++ cmd/frps_allowed_ports/ports
@@ -1,0 +1,6 @@
+user1=6000
+user1=321
+user1=6262
+user1=service
+user1=service.masternetwork.us
+user2=124
