+++ pkg/server/controller/op.go
@@ -1,21 +1,19 @@
 package controller
 
 import (
-	"net/http"
 	"fmt"
+	"net/http"
+	"strings"
+
 	plugin "github.com/fatedier/frp/pkg/plugin/server"
 	"github.com/gin-gonic/gin"
 
-  	 "strconv"
-
+	"strconv"
 )
-
-
 
 type OpController struct {
 	ports map[string][]string
 }
-
 
 func NewOpController(ports map[string][]string) *OpController {
 	return &OpController{
@@ -32,19 +30,13 @@ func contains(s []string, str string) bool {
 			return true
 		}
 	}
-
 	return false
 }
 
 func (c *OpController) HandleLogin(ctx *gin.Context) (interface{}, error) {
-
-	var r plugin.Request 
-
+	var r plugin.Request
 	var content plugin.NewProxyContent
-	
 	var res plugin.Response
-
-	var find bool
 
 	r.Content = &content
 	if err := ctx.BindJSON(&r); err != nil {
@@ -53,50 +45,45 @@ func (c *OpController) HandleLogin(ctx *gin.Context) (interface{}, error) {
 			Err:  err,
 		}
 	}
-	
 
-	fmt.Println("---------------------------------")
-	fmt.Println(content.ProxyName)
-	fmt.Println(content.ProxyType)
-	fmt.Println(content.RemotePort)
-	fmt.Println(content.CustomDomains)
-	fmt.Println(content.SubDomain)
-	fmt.Println("---------------------------------")
+	fmt.Println("-------------Plugin: Allowed Ports--------------------")
+	fmt.Printf("ProxyName: %s\tProxyType%s\t", content.ProxyName, content.ProxyType)
+	if strings.ToLower(content.ProxyType) == "tcp" || strings.ToLower(content.ProxyType) == "udp" {
+		fmt.Printf("RemotePort: %d\r\n", content.RemotePort)
+	} else if strings.HasPrefix(content.ProxyType, "http") {
+		fmt.Printf("CustomDomains%s\r\n", content.CustomDomains)
+	} else {
+		fmt.Println("Won't do validation for this type")
+		res.Unchange = true
+		return res, nil
+	}
 
-	subdomain  := content.SubDomain
+	subdomain := content.SubDomain
 	remoteport := strconv.Itoa(content.RemotePort)
 	username := content.User.User
 
-
-	
-	if subdomain == "" && remoteport == "0" && len(content.CustomDomains) == 0{
-		fmt.Println("Rejected")
+	if subdomain == "" && remoteport == "0" && len(content.CustomDomains) == 0 {
 		res.Reject = true
-		res.RejectReason = "Misconfiguration of the client"
+		res.RejectReason = "Rejected due to misconfiguration of the client"
 	}
 
-	find = false
+	find := false
 
 	for _, port_allowed := range c.ports[username] {
-	  if port_allowed == remoteport || port_allowed == subdomain  {
-		find = true
+		if port_allowed == remoteport || port_allowed == subdomain {
+			find = true
+		}
 
+		if contains(content.CustomDomains, port_allowed) {
+			find = true
+		}
 	}
-
-	if contains(content.CustomDomains, port_allowed) {
-		find = true
-
-	}
-    }
-	if find == false {
-	
-		fmt.Println("Client is not allowed to use this port")
+	if !find {
 		res.Reject = true
-		res.RejectReason = "Not allowed => Port or subdomain false"
-		
+		res.RejectReason = "Client is not allowed => Port or subdomain false"
 	}
-	
-	if res.Reject != true {
+
+	if !res.Reject {
 		res.Unchange = true
 	}
 	return res, nil
