--- releases/Alpha/ports
@@ -1,3 +1,0 @@
-user1=123
-user1=321
-user2=124
